complete 4.21 bloglist expansion, step 9

+++ part4/bloglist/controller/blogs.js
@@ -35,7 +35,6 @@ blogRouter.post("/", async (request, response, next) => {
 
   try {
     const savedPost = await blogPost.save();
-    console.log(savedPost);
     user.blogs = user.blogs.concat(savedPost._id);
     await user.save();
     response.status(201).json(savedPost);
@@ -45,8 +44,26 @@ blogRouter.post("/", async (request, response, next) => {
 });
 
 blogRouter.delete("/:id", async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid " });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  // console.log(blog.user._id.toString() === user.id.toString());
+  if (blog.user._id.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: "invalid action" });
+  }
+
   try {
     await Blog.findByIdAndRemove(request.params.id);
+    user.blogs = user.blogs.filter(
+      (blog) => blog.toString() !== request.params.id
+    );
+    await user.save();
     response.status(204).end();
   } catch (exception) {
     next(exception);
